seperating components into different files to keep things modular
