+++ hacker-stories/src/Item.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Item = ({ title, url, director, releaseYear }) => {
+	return (
+		<li>
+			<span>
+				<a href={url}>{title} </a>
+			</span>
+			<span> was directed by {director}</span>
+			<span> and released in {releaseYear}</span>
+		</li>
+	)
+}
+
+export default Item;
